requirements.txt
streamlit==2.0.0
streamlit-option-menu==0.3.6
numpy==1.24.4
pandas==2.0.3
plotly==5.15.0
pillow==10.0.1
torch==2.1.0
torchvision==0.16.0
timm==0.9.5
openai==1.3.6
opencv-python==4.8.1.78
pydicom==2.4.3
scikit-image==0.21.0
tqdm==4.66.1
seaborn==0.12.2
scikit-learn==1.3.1
scipy==1.11.2
httpx==0.28.1
